fix(naming): fix reducers name
- fix reducers name
- fix actions names
