--- src/actions/ViewMenu/actionTypes.js
@@ -1,10 +1,0 @@
-export const VIEW_MENU_START = 'VIEW_MENU_START';
-export const VIEW_MENU_SUCCESS = 'VIEW_MENU_SUCCESS';
-export const VIEW_MENU_FAIL = 'VIEW_MENU_FAIL';
-
-
-export const POST_MENU_START = 'POST_MENU_START';
-export const POST_MENU_SUCCESS = 'POST_MENU_SUCCESS';
-export const POST_MENU_FAIL = 'POST_MENU_FAIL';
-
-export const UPDATE_CART_ICON = 'UPDATE_CART_ICON';
