+++ src/actions/menu/actionTypes.js
@@ -1,0 +1,10 @@
+export const VIEW_MENU_START = 'VIEW_MENU_START';
+export const VIEW_MENU_SUCCESS = 'VIEW_MENU_SUCCESS';
+export const VIEW_MENU_FAIL = 'VIEW_MENU_FAIL';
+
+
+export const PLACE_ORDER_START = 'PLACE_ORDER_START';
+export const PLACE_ORDER_SUCCESS = 'POST_MENU_SUCCESS';
+export const PLACE_ORDER_FAIL = 'POST_MENU_FAIL';
+
+export const UPDATE_CART = 'UPDATE_CART';
